Delete LLMDatasetMessage in favour of LLMMessage with limited to_dict/from_dict
